[THA-56] Remove user entity and signup endpoint. Connect service with bff (#57)
* [THA-56] Remove user entity related stuff and replace them with a userId coming from the auth service

* [THA-56] Update validation annotations

* [THA-56] Update remove booking operation

* [THA-56] Add transactional annotation and null checks

* [THA-56] Fix bugs in hotel client

* [THA-56] Add transactional annotation and refactor

* [THA-56] Update controller

* [THA-56] Move base processor to core module

* [THA-56] Create Output interface

+++ core/src/main/java/com/tinqinacademy/hotel/core/processors/deleteroom/DeleteRoomOperationProcessor.java
@@ -1,11 +1,13 @@
 package com.tinqinacademy.hotel.core.processors.deleteroom;
 
-import com.tinqinacademy.hotel.api.base.BaseOperationProcessor;
+import com.tinqinacademy.hotel.core.processors.base.BaseOperationProcessor;
 import com.tinqinacademy.hotel.api.errors.ErrorOutput;
 import com.tinqinacademy.hotel.api.exceptions.EntityNotFoundException;
+import com.tinqinacademy.hotel.api.exceptions.RoomUnavailableException;
 import com.tinqinacademy.hotel.api.operations.deleteroom.input.DeleteRoomInput;
 import com.tinqinacademy.hotel.api.operations.deleteroom.operation.DeleteRoomOperation;
 import com.tinqinacademy.hotel.api.operations.deleteroom.output.DeleteRoomOutput;
+import com.tinqinacademy.hotel.persistence.entities.booking.Booking;
 import com.tinqinacademy.hotel.persistence.entities.room.Room;
 import com.tinqinacademy.hotel.persistence.repositories.BookingRepository;
 import com.tinqinacademy.hotel.persistence.repositories.RoomRepository;
@@ -16,7 +18,10 @@ import org.springframework.core.convert.ConversionService;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import jakarta.validation.Validator;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
+import java.util.List;
 import java.util.UUID;
 
 import static io.vavr.API.Match;
@@ -35,32 +40,35 @@ public class DeleteRoomOperationProcessor extends BaseOperationProcessor impleme
   }
 
   @Override
+  @Transactional
   public Either<ErrorOutput, DeleteRoomOutput> process(DeleteRoomInput input) {
     return validateInput(input)
         .flatMap(validInput ->
-                Try.of(() -> {
-                      log.info("Start deleteRoom input: {}", validInput);
-                      UUID roomId = UUID.fromString(input.getId());
+            Try.of(() -> {
+                  log.info("Start deleteRoom input: {}", validInput);
+                  UUID roomId = UUID.fromString(input.getId());
 
-                      Room room = getRoomByIdOrThrow(roomId);
+                  Room room = getRoomByIdOrThrow(roomId);
+                  List<Booking> roomBookings = bookingRepository.getBookingsOfRoomForPeriod(roomId, LocalDate.now(), LocalDate.now().plusYears(10_000));
 
-//      if (!roomBookings.isEmpty()) {
-//        throw new RoomUnavailableException("Unable to delete room: Room is still being used");
-//      }
+                  if (!roomBookings.isEmpty()) {
+                    throw new RoomUnavailableException("Unable to delete room: Room is still being used");
+                  }
 
-                      bookingRepository.deleteBookingsByRoom(room);
-                      roomRepository.delete(room);
+                  bookingRepository.deleteBookingsByRoom(room);
+                  roomRepository.delete(room);
 
-                      DeleteRoomOutput output = createOutput();
-                      log.info("End deleteRoom output: {}", output);
-                      return output;
-                    })
-                    .toEither()
-                    .mapLeft(t -> Match(t).of(
-                        customStatusCase(t, EntityNotFoundException.class, HttpStatus.NOT_FOUND),
-                        customStatusCase(t, IllegalArgumentException.class, HttpStatus.UNPROCESSABLE_ENTITY),
-                        defaultCase(t)
-                    ))
+                  DeleteRoomOutput output = createOutput();
+                  log.info("End deleteRoom output: {}", output);
+                  return output;
+                })
+                .toEither()
+                .mapLeft(t -> Match(t).of(
+                    customStatusCase(t, EntityNotFoundException.class, HttpStatus.NOT_FOUND),
+                    customStatusCase(t, RoomUnavailableException.class, HttpStatus.CONFLICT),
+                    customStatusCase(t, IllegalArgumentException.class, HttpStatus.UNPROCESSABLE_ENTITY),
+                    defaultCase(t)
+                ))
         );
   }
